完成限流规则CRUD功能

+++ src/views/ratelimit/index.vue
@@ -1,12 +1,13 @@
 <template>
   <div>
-    <Modal :title="'abc'" ref="modal"></Modal>
+    <Modal ref="modal" :formViewName="'RateLimitEdit'" @onSave="save"></Modal>
     <div
       class="d-flex justify-content-between flex-wrap flex-md-nowrap align-items-center pt-3 pb-2 mb-3 border-bottom">
       <h1 class="h4">{{$t('menu.rateLimit')}}</h1>
       <div class="btn-toolbar mb-2 mb-md-0">
         <div class="btn-group me-2">
-          <button type="button" class="btn btn-sm btn-outline-primary" @click="showModal">
+          <button type="button" class="btn btn-sm btn-outline-primary"
+                  @click="showModal($t('rateLimit.modal.title'),{},false)">
             {{$t('button.add')}}
           </button>
         </div>
@@ -14,11 +15,15 @@
     </div>
     <form class="row col-3">
       <div class="input-group mb-3">
-        <span class="input-group-text">{{$t('commons.keyword')}}</span>
+        <div class="input-group-prepend">
+          <span class="input-group-text">{{$t('commons.keyword')}}</span>
+        </div>
         <input type="text" class="form-control" :placeholder='$t("rateLimit.input.keyword")'
                v-model="queryObject.keywords">
-        <button class="btn btn-outline-primary" type="button" @click="queryRateLimits">{{$t('commons.search')}}
-        </button>
+        <div class="input-group-append">
+          <button class="btn btn-outline-primary" type="button" @click="queryRateLimits">{{$t('commons.search')}}
+          </button>
+        </div>
       </div>
     </form>
     <div class="table-responsive">
@@ -44,15 +49,18 @@
           <td>{{item.frequency}}</td>
           <td>{{item.dimension}}</td>
           <td>
-            <span class="badge"
-                  :class="{'bg-success':item.enable,'bg-danger':!item.enable}">
-            {{item.enable? $t('commons.enabled'):$t('commons.disabled')}}</span>
+            <span :class="{'text-success':item.enable,'text-danger':!item.enable}">
+                   {{item.enable? $t('commons.enabled'):$t('commons.disabled')}}
+           </span>
           </td>
           <td>{{item.createTime | dateFormat}}</td>
           <td>{{item.modifyTime | dateFormat}}</td>
           <td>
             <div class="btn-group btn-group-sm" role="group">
-              <button class="btn btn btn-outline-primary">{{$t('commons.edit')}}</button>
+              <button class="btn btn btn-outline-success"
+                      @click="showModal($t('rateLimit.modal.title'),item, true)">
+                {{$t('commons.edit')}}
+              </button>
               <button class="btn btn btn-outline-danger" @click="deleteRecord(item.id)">{{$t('commons.delete')}}
               </button>
             </div>
@@ -67,6 +75,7 @@
 <script>
   import {RateLimitService} from "@/service";
   import Modal from "../../components/Modal";
+  import edit from "./edit";
 
   export default {
     name: "rateLimit",
@@ -83,13 +92,41 @@
         },
       }
     },
-    components: {Modal},
+    components: {Modal, edit},
     mounted() {
       this.queryRateLimits();
     },
     methods: {
-      showModal(){
-        this.$refs.modal.modalToggle();
+      save(data) {
+        if (data.editFlag) {
+          this.editRecord(data.data)
+        } else {
+          this.addRecord(data.data)
+        }
+      },
+      showModal(title, data, editFlag) {
+        const that = this;
+        if (Object.keys(data).length === 0
+          && data.constructor === Object) {
+          data = {
+            "scope": "APPLICATION",
+            "dimension": "MINUTE",
+            "frequency": 1,
+            "enable": true
+          }
+        }
+        this.$refs.modal.modalToggle({
+          title: title,
+          data: data,
+          editFlag: editFlag,
+          saveFunction: () => {
+            if (editFlag) {
+              that.editRecord();
+            } else {
+              that.addRecord();
+            }
+          }
+        });
       },
       queryRateLimits() {
         const that = this;
@@ -97,11 +134,34 @@
           that.pageResult = response.data.data;
         })
       },
+      addRecord(item) {
+        const that = this;
+        RateLimitService.addRecord(item).then(response => {
+          if (response.data.success) {
+            that.queryRateLimits();
+          } else {
+            alert(response.data.message)
+          }
+        })
+      },
+      editRecord(item) {
+        const that = this;
+        RateLimitService.editRecord(item).then(response => {
+          if (response.data.success) {
+            that.queryRateLimits();
+          } else {
+            alert(response.data.message)
+          }
+        })
+      },
       deleteRecord(id) {
         const that = this;
-        RateLimitService.deleteRecord(id).then(response => {
-          that.queryRateLimits();
-        })
+        //FIXME confirm
+        if (confirm("delete a record? ")) {
+          RateLimitService.deleteRecord(id).then(response => {
+            that.queryRateLimits();
+          })
+        }
       }
     }
   }
